Extracted EventArgs.cs into separate files.
