+++ SevenZip/EventArguments/ExtractFileCallback.cs
@@ -1,0 +1,11 @@
+﻿#if UNMANAGED
+
+namespace SevenZip
+{
+    /// <summary>
+    /// Callback delegate for <see cref="SevenZipExtractor.ExtractFiles(SevenZip.ExtractFileCallback)"/>.
+    /// </summary>
+    public delegate void ExtractFileCallback(ExtractFileCallbackArgs extractFileCallbackArgs);
+}
+
+#endif
