+++ SevenZip/LZMA/LzmaEncodeStream.cs
