+++ SevenZip/LZMA/LzmaProgressCallback.cs
