+++ SevenZip/SevenZip.csproj
@@ -87,7 +87,16 @@
     <Compile Include="ArchiveOpenCallback.cs" />
     <Compile Include="ArchiveUpdateCallback.cs" />
     <Compile Include="Com.cs" />
-    <Compile Include="EventArgs.cs" />
+    <Compile Include="EventArguments\ExtractFileCallbackArgs.cs" />
+    <Compile Include="EventArguments\ExtractFileCallbackReason.cs" />
+    <Compile Include="EventArguments\FileInfoEventArgs.cs" />
+    <Compile Include="EventArguments\FileNameEventArgs.cs" />
+    <Compile Include="EventArguments\FileOverwriteEventArgs.cs" />
+    <Compile Include="EventArguments\ICancellable.cs" />
+    <Compile Include="EventArguments\IntEventArgs.cs" />
+    <Compile Include="EventArguments\OpenEventArgs.cs" />
+    <Compile Include="EventArguments\PercentDoneEventArgs.cs" />
+    <Compile Include="EventArguments\ProgressEventArgs.cs" />
     <Compile Include="Exceptions\CompressionFailedException.cs" />
     <Compile Include="Exceptions\ExtractionFailedException.cs" />
     <Compile Include="Exceptions\LzmaException.cs" />
@@ -98,14 +107,15 @@
     <Compile Include="Exceptions\SevenZipInvalidFileNamesException.cs" />
     <Compile Include="Exceptions\SevenZipLibraryException.cs" />
     <Compile Include="Exceptions\SevenZipSfxValidationException.cs" />
+    <Compile Include="EventArguments\ExtractFileCallback.cs" />
     <Compile Include="FileChecker.cs" />
     <Compile Include="Formats.cs" />
     <Compile Include="LibraryFeature.cs" />
     <Compile Include="LibraryManager.cs" />
     <Compile Include="Common.cs" />
-    <Compile Include="LzmaDecodeStream.cs" />
-    <Compile Include="LzmaEncodeStream.cs" />
-    <Compile Include="LzmaProgressCallback.cs" />
+    <Compile Include="LZMA\LzmaDecodeStream.cs" />
+    <Compile Include="LZMA\LzmaEncodeStream.cs" />
+    <Compile Include="LZMA\LzmaProgressCallback.cs" />
     <Compile Include="NativeMethods.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="sdk\Common\CRC.cs" />
